storage: extract closedts code from replica
Release note: None
